Refactored chat rooms page (former friends page) - need to readd chat link once it's refactored
